change scors to ints

+++ Models/Rating.cs
@@ -8,9 +8,9 @@
         private Movie _movie;
         public Movie Movie { get { return _movie; } }   
 
-        private double _score;
-        public double Score { get { return _score; } }    
-        public Rating(User user, Movie movie, double score) 
+        private int _score;
+        public int Score { get { return _score; } }    
+        public Rating(User user, Movie movie, int score) 
         {
             _user = user;
             _movie = movie;
